+++ frontend/src/hooks/useIsLogin.ts
@@ -1,9 +1,9 @@
-import { useIsLoginQuery } from "../api/auth-api";
-import { useAppSelector } from "./hooks";
+import { useIsLoginQuery } from '../api/auth-api/auth-api';
+import { useAppSelector } from './hooks';
 
 export const useIsLogin = () => {
-    const { isLoading } = useIsLoginQuery();
-    const isLogin = useAppSelector(state => state.auth.isLogin);
+  const { isLoading } = useIsLoginQuery();
+  const isLogin = useAppSelector((state) => state.auth.isLogin);
 
-    return { isLoading, isLogin }
-}
+  return { isLoading, isLogin };
+};