+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Project utils
 
+- Add redux query validation schemas 
+
 - Add separate /backend/.env.prod file for a production build
 
 <--------------------------